Move early lint machanism in librustc_lint.
