+++ src/librustc_lint/lib.rs
@@ -24,6 +24,7 @@ extern crate rustc_session;
 
 mod array_into_iter;
 pub mod builtin;
+mod early;
 mod non_ascii_idents;
 mod nonstandard_style;
 mod redundant_semicolon;
@@ -57,6 +58,7 @@ use unused::*;
 
 /// Useful for other parts of the compiler.
 pub use builtin::SoftLints;
+pub use early::check_ast_crate;
 
 pub fn provide(providers: &mut Providers<'_>) {
     *providers = Providers { lint_mod, ..*providers };
